refactor: relocate reviewer warnings and consolidate shared helpers
- Move PR advisory formatting to `comment-builder/reviewer-warnings.util`
- Reuse frontmatter and heading guard detection in PR details
- Add `createGuardLabelResolver` for `pt-br` and `ru` advisory labels
- Extract `markdown-path`, `github-url`, and `rate-limit-detect` from `common.util`
