+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Format: [Keep a Changelog](https://keepachangelog.com/en/1.1.0/). Versioning: [S
 ### Changed
 
 - Legacy full-body and chunked markdown translation now live in a dedicated service; segment-batch failures no longer fall back on auth, quota, or non-splittable errors.
+- Advisory validation PR details reuse the same frontmatter and heading detection as post-translation guards instead of parallel regex logic.
+- Locale advisory guard labels share one resolver factory across `pt-br` and `ru`.
+- PR reviewer-warning formatting now lives under `comment-builder` with the other GitHub comment utilities.
+- Shared app utilities split markdown path filtering, GitHub URL builders, and rate-limit detection into topic modules while keeping `@/app/utils` exports stable.
 - Fork synchronization uses the fork repository default branch instead of assuming `main`.
 - Concurrent `translateContent` calls on the shared translator service no longer share per-file LLM usage or translation-path state.
 - Language-detector tag stripping and JSX static-text link analyzers use linear scans instead of nested-regex patterns that could backtrack on near-valid upstream markdown.
