+++ src/app/services/comment-builder/index.ts
@@ -1,2 +1,3 @@
 export * from "./comment-builder.service";
 export * from "./progress-comment.util";
+export * from "./reviewer-warnings.util";
