+++ src/app/services/translator/validation/guards/frontmatter-preserved.guard.ts
@@ -3,6 +3,38 @@ import type { PostTranslationValidationGuard } from "../validation.types";
 import { MARKDOWN_REGEXES } from "../../markdown/markdown.regexes";
 import { POST_TRANSLATION_GUARD_IDS } from "../validation.constants";
 
+/** Frontmatter block dropped from a translation that had YAML frontmatter in the source */
+export interface FrontmatterPreservedViolation {
+	readonly sourceFrontmatterBlockLength: number;
+}
+
+/**
+ * Detects when the source has YAML frontmatter but the translation dropped it.
+ *
+ * @param source Original markdown before translation
+ * @param translated Model output to validate
+ *
+ * @returns Violation metadata, or `null` when frontmatter is preserved or absent in the source
+ */
+export function detectFrontmatterPreservedViolation(
+	source: string,
+	translated: string,
+): FrontmatterPreservedViolation | null {
+	const sourceMatch = MARKDOWN_REGEXES.frontmatter.exec(source);
+	MARKDOWN_REGEXES.frontmatter.lastIndex = 0;
+	if (!sourceMatch?.groups?.["content"]) {
+		return null;
+	}
+
+	const translatedMatch = MARKDOWN_REGEXES.frontmatter.exec(translated)?.groups?.["content"];
+	MARKDOWN_REGEXES.frontmatter.lastIndex = 0;
+	if (translatedMatch) {
+		return null;
+	}
+
+	return { sourceFrontmatterBlockLength: sourceMatch[0].length };
+}
+
 /**
  * Rejects translations that dropped YAML frontmatter
  *
@@ -12,15 +44,8 @@ import { POST_TRANSLATION_GUARD_IDS } from "../validation.constants";
  * @returns Guard failure with retry hint, or `null` when frontmatter is preserved
  */
 export const frontmatterPreservedGuard: PostTranslationValidationGuard = (source, translated) => {
-	const originalMatch = MARKDOWN_REGEXES.frontmatter.exec(source)?.groups?.["content"];
-	if (!originalMatch) return null;
-
-	MARKDOWN_REGEXES.frontmatter.lastIndex = 0;
-
-	const translatedMatch = MARKDOWN_REGEXES.frontmatter.exec(translated)?.groups?.["content"];
-	MARKDOWN_REGEXES.frontmatter.lastIndex = 0;
-
-	if (translatedMatch) return null;
+	const violation = detectFrontmatterPreservedViolation(source, translated);
+	if (!violation) return null;
 
 	return {
 		guardId: POST_TRANSLATION_GUARD_IDS.frontmatterPreserved,
